Split components into their own files
